+++ sdk/resourcemanager/containerservice/armcontainerservice/CHANGELOG.md
@@ -1,5 +1,101 @@
 # Release History
 
+## 2.0.0-beta.1 (2022-06-02)
+### Breaking Changes
+
+- Struct `ManagedClusterSecurityProfileAzureDefender` has been removed
+- Field `AzureDefender` of struct `ManagedClusterSecurityProfile` has been removed
+
+### Features Added
+
+- New const `KeyVaultNetworkAccessTypesPrivate`
+- New const `TrustedAccessRoleBindingProvisioningStateUpdating`
+- New const `TrustedAccessRoleBindingProvisioningStateDeleting`
+- New const `SnapshotTypeManagedCluster`
+- New const `TrustedAccessRoleBindingProvisioningStateSucceeded`
+- New const `KeyVaultNetworkAccessTypesPublic`
+- New const `NetworkPluginNone`
+- New const `NetworkPluginModeOverlay`
+- New const `TrustedAccessRoleBindingProvisioningStateFailed`
+- New const `OSSKUWindows2019`
+- New const `OSSKUWindows2022`
+- New function `PossibleKeyVaultNetworkAccessTypesValues() []KeyVaultNetworkAccessTypes`
+- New function `PossibleNetworkPluginModeValues() []NetworkPluginMode`
+- New function `TrustedAccessRoleBindingProperties.MarshalJSON() ([]byte, error)`
+- New function `*ManagedClustersClient.BeginRotateServiceAccountSigningKeys(context.Context, string, string, *ManagedClustersClientBeginRotateServiceAccountSigningKeysOptions) (*runtime.Poller[ManagedClustersClientRotateServiceAccountSigningKeysResponse], error)`
+- New function `ManagedClusterSnapshot.MarshalJSON() ([]byte, error)`
+- New function `PossibleTrustedAccessRoleBindingProvisioningStateValues() []TrustedAccessRoleBindingProvisioningState`
+- New struct `AzureKeyVaultKms`
+- New struct `ManagedClusterIngressProfile`
+- New struct `ManagedClusterIngressProfileWebAppRouting`
+- New struct `ManagedClusterOIDCIssuerProfile`
+- New struct `ManagedClusterPropertiesForSnapshot`
+- New struct `ManagedClusterSecurityProfileDefender`
+- New struct `ManagedClusterSecurityProfileDefenderSecurityMonitoring`
+- New struct `ManagedClusterSecurityProfileWorkloadIdentity`
+- New struct `ManagedClusterSnapshot`
+- New struct `ManagedClusterSnapshotListResult`
+- New struct `ManagedClusterSnapshotProperties`
+- New struct `ManagedClusterSnapshotsClientCreateOrUpdateOptions`
+- New struct `ManagedClusterSnapshotsClientCreateOrUpdateResponse`
+- New struct `ManagedClusterSnapshotsClientDeleteOptions`
+- New struct `ManagedClusterSnapshotsClientDeleteResponse`
+- New struct `ManagedClusterSnapshotsClientGetOptions`
+- New struct `ManagedClusterSnapshotsClientGetResponse`
+- New struct `ManagedClusterSnapshotsClientListByResourceGroupOptions`
+- New struct `ManagedClusterSnapshotsClientListByResourceGroupResponse`
+- New struct `ManagedClusterSnapshotsClientListOptions`
+- New struct `ManagedClusterSnapshotsClientListResponse`
+- New struct `ManagedClusterSnapshotsClientUpdateTagsOptions`
+- New struct `ManagedClusterSnapshotsClientUpdateTagsResponse`
+- New struct `ManagedClusterStorageProfileBlobCSIDriver`
+- New struct `ManagedClusterWorkloadAutoScalerProfile`
+- New struct `ManagedClusterWorkloadAutoScalerProfileKeda`
+- New struct `ManagedClustersClientBeginRotateServiceAccountSigningKeysOptions`
+- New struct `ManagedClustersClientRotateServiceAccountSigningKeysResponse`
+- New struct `NetworkProfileForSnapshot`
+- New struct `TrustedAccessRole`
+- New struct `TrustedAccessRoleBinding`
+- New struct `TrustedAccessRoleBindingListResult`
+- New struct `TrustedAccessRoleBindingProperties`
+- New struct `TrustedAccessRoleBindingsClientCreateOrUpdateOptions`
+- New struct `TrustedAccessRoleBindingsClientCreateOrUpdateResponse`
+- New struct `TrustedAccessRoleBindingsClientDeleteOptions`
+- New struct `TrustedAccessRoleBindingsClientDeleteResponse`
+- New struct `TrustedAccessRoleBindingsClientGetOptions`
+- New struct `TrustedAccessRoleBindingsClientGetResponse`
+- New struct `TrustedAccessRoleBindingsClientListOptions`
+- New struct `TrustedAccessRoleBindingsClientListResponse`
+- New struct `TrustedAccessRoleListResult`
+- New struct `TrustedAccessRoleRule`
+- New struct `TrustedAccessRolesClientListOptions`
+- New struct `TrustedAccessRolesClientListResponse`
+- New field `MessageOfTheDay` in struct `ManagedClusterAgentPoolProfile`
+- New field `EnableCustomCATrust` in struct `ManagedClusterAgentPoolProfile`
+- New field `HostGroupID` in struct `ManagedClusterAgentPoolProfile`
+- New field `CapacityReservationGroupID` in struct `ManagedClusterAgentPoolProfile`
+- New field `WorkloadIdentity` in struct `ManagedClusterSecurityProfile`
+- New field `AzureKeyVaultKms` in struct `ManagedClusterSecurityProfile`
+- New field `Defender` in struct `ManagedClusterSecurityProfile`
+- New field `BlobCSIDriver` in struct `ManagedClusterStorageProfile`
+- New field `IgnorePodDisruptionBudget` in struct `AgentPoolsClientBeginDeleteOptions`
+- New field `IgnorePodDisruptionBudget` in struct `ManagedClustersClientBeginDeleteOptions`
+- New field `EnableNamespaceResources` in struct `ManagedClusterProperties`
+- New field `WorkloadAutoScalerProfile` in struct `ManagedClusterProperties`
+- New field `CreationData` in struct `ManagedClusterProperties`
+- New field `IngressProfile` in struct `ManagedClusterProperties`
+- New field `OidcIssuerProfile` in struct `ManagedClusterProperties`
+- New field `HostGroupID` in struct `ManagedClusterAgentPoolProfileProperties`
+- New field `CapacityReservationGroupID` in struct `ManagedClusterAgentPoolProfileProperties`
+- New field `EnableCustomCATrust` in struct `ManagedClusterAgentPoolProfileProperties`
+- New field `MessageOfTheDay` in struct `ManagedClusterAgentPoolProfileProperties`
+- New field `Version` in struct `ManagedClusterStorageProfileDiskCSIDriver`
+- New field `EnableVnetIntegration` in struct `ManagedClusterAPIServerAccessProfile`
+- New field `SubnetID` in struct `ManagedClusterAPIServerAccessProfile`
+- New field `NetworkPluginMode` in struct `NetworkProfile`
+- New field `EffectiveNoProxy` in struct `ManagedClusterHTTPProxyConfig`
+
+
 ## 1.0.0 (2022-05-16)
 
 The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html) since version 1.0.0, which contains breaking changes.